Improved prototyping of loading data
