--- DBTest/DBTest.sln
@@ -1,16 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DBTest", "DBTest\DBTest.csproj", "{BAE66932-5D2F-4C84-83A9-D74B25A93C89}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{BAE66932-5D2F-4C84-83A9-D74B25A93C89}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{BAE66932-5D2F-4C84-83A9-D74B25A93C89}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{BAE66932-5D2F-4C84-83A9-D74B25A93C89}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{BAE66932-5D2F-4C84-83A9-D74B25A93C89}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-EndGlobal
